[type:feature] ai proxy plugin
