+++ shenyu-common/src/main/java/org/apache/shenyu/common/enums/AiModelProviderEnum.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shenyu.common.enums;
+
+/**
+ * The Ai model provider enum.
+ */
+public enum AiModelProviderEnum {
+    
+    /**
+     * OpenAI.
+     */
+    OPEN_AI("OpenAI"),
+    
+    /**
+     * DeepSeek.
+     */
+    DEEP_SEEK("DeepSeek"),
+    
+    /**
+     * Ali.
+     */
+    A_LI("ALi"),
+    
+    ;
+
+    private final String name;
+
+    AiModelProviderEnum(final String name) {
+        this.name = name;
+    }
+
+    /**
+     * get name.
+     *
+     * @return name
+     */
+    public String getName() {
+        return name;
+    }
+}
