+++ shenyu-plugin/shenyu-plugin-ai-proxy/src/main/java/org/apache/shenyu/plugin/ai/proxy/strategy/AiModelFactory.java
@@ -1,0 +1,53 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shenyu.plugin.ai.proxy.strategy;
+
+import org.apache.shenyu.plugin.ai.proxy.strategy.ali.Ali;
+import org.apache.shenyu.plugin.ai.proxy.strategy.deepseek.DeepSeek;
+import org.apache.shenyu.plugin.ai.proxy.strategy.openai.OpenAI;
+
+/**
+ * The interface Ai model.
+ */
+public final class AiModelFactory {
+    
+    private AiModelFactory() {
+    }
+
+    /**
+     * Create ai model.
+     *
+     * @param provider the provider
+     * @return the ai model
+     */
+    public static AiModel createAiModel(final String provider) {
+        if (provider == null) {
+            return null;
+        }
+        switch (provider) {
+            case "OpenAI":
+                return new OpenAI();
+            case "DeepSeek":
+                return new DeepSeek();
+            case "ALi":
+                return new Ali();
+            default:
+                return null;
+        }
+    }
+}
